requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.3
PyMuPDF==1.23.7
faiss-cpu==1.11.0.post1
sentence-transformers==5.0.0
beautifulsoup4==4.13.0
scikit-learn==1.7.1
python-docx==1.2.0
